Introduce module `rx_replace.py`
It contains a base regex replacer, type declaration and helper function.
